+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 ## [Unreleased]
 ### Added
 - Etherscan parser: added internal transactions export.
+- Binance parser: added cash deposit and withdrawal exports.
+- Binance parser: added statements export.
 ### Changed
 - Conversion tool: UnknownAddressError exception changed to generic DataFilenameError.
+- Binance parser: use filename to determine if deposits or withdrawals.
 
 ## Version [0.4.3] Beta (2020-12-04)
 Important:- if upgrading, please remove your price data cache file for CryptoCompare: `~/.bittytax/cache/CryptoCompare.json` (see Issue [#29](https://github.com/BittyTax/BittyTax/issues/29))